\name{liureg-package}
\alias{liureg}
\alias{liureg-package}
\title{Liu Regression and Estimator}
\description{R package for fitting linear Liu Regression and Estimator.}

\details{
This package contains functions for fitting Liu regression models, including function for computation of different Liu related statistics (such as MSE, Var-Cov matrix, R-squared), estimation of biasing parameter from different researchers, testing of Liu coefficients, model selection criteria, residuals, predicted values and fitted values. The package also includes function for plotting of Liu coefficients and different Liu statistics for selection of optimal value of biasing parameter \eqn{d}{d}.

For a complete list of functions, use \code{library(help="liureg")}.
}

\author{Muahmmad Imdadullah, Dr. Muhammad Aslam}
